pyproject.toml
[tool.poetry]
name = "python-fb-page-insights-client"
version = "0.1.0"
description = ""
authors = ["Grimmer <grimmer0125@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
pydantic = "^1.8.2"
requests = "^2.25.1"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
